fix:修改資料庫結構
1.調整DB結構
2.修改部份宣告
3.修改會計師群組管理(尚未完成)

+++ SealTypographicWebAPI/Models/PaginateSearch.cs
@@ -21,7 +21,7 @@ namespace SealTypographicWebAPI.Models
         /// </summary>
         /// <example>5</example>
         [Required]
-        [Range(1, PageSizeLimit.Max)]
+        [Range(1, (int)PageSizeLimit.Max)]
         public int PageSize { get; set; }
     }
